Add files via upload
This project involves simulating a system , using DHT22 sensors for temperature and humidity, a light sensor, and a slider for CO2 levels. It will implement scenarios like opening windows with a servo motor for high CO2 and adjusting heating or cooling based on temperature. Data will be displayed on ThingSpeak for real-time monitoring and trend analysis of indoor conditions.